requirements.txt
asgiref==3.6.0
Django==4.1.6
django-crum==0.7.9
django-jazzmin==2.6.0
gunicorn==20.1.0
python-dotenv==0.21.1
sqlparse==0.4.3
